+++ index.php
@@ -69,17 +69,21 @@ require 'vendor/autoload.php';
 
 require('Guide/overview.php');
 require('Guide/alert.php');
+require('Guide/badge.php');
 require('Guide/button.php');
 require('Guide/button_group.php');
 require('Guide/button_toolbar.php');
 require('Guide/dropdown.php');
 require('Guide/icon.php');
+require('Guide/image.php');
+require('Guide/label.php');
 require('Guide/modal.php');
 require('Guide/nav.php');
 require('Guide/navbar.php');
 require('Guide/panel.php');
 require('Guide/panel_rows.php');
 require('Guide/progress_bar.php');
+require('Guide/thumbnail.php');
 
 use cobisja\BootHelp\Nav;
 use cobisja\BootHelp\Helpers\LinkTo;
@@ -92,17 +96,21 @@ use cobisja\BootHelp\Helpers\LinkTo;
                             return [
                                 new LinkTo('Overview', ['href'=>'#overview']),
                                 new LinkTo('Alerts', ['href'=>'#alerts']),
+                                new LinkTo('Badges', ['href'=>'#badges']),
                                 new LinkTo('Buttons', ['href'=>'#buttons']),
                                 new LinkTo('Button Group', ['href'=>'#button-group']),
                                 new LinkTo('Button Toolbar', ['href'=>'#button-toolbar']),
                                 new LinkTo('Dropdowns', ['href'=>'#dropdowns']),
                                 new LinkTo('Icons', ['href'=>'#icons']),
+                                new LinkTo('Images', ['href'=>'#images']),
+                                new LinkTo('Labels', ['href'=>'#labels']),
                                 new LinkTo('Modals', ['href'=>'#modals']),
                                 new LinkTo('Navs', ['href'=>'#navs']),
                                 new LinkTo('Navbars', ['href'=>'#navbars']),
                                 new LinkTo('Panels', ['href'=>'#panels']),
                                 new LinkTo('Panel rows', ['href'=>'#panel-rows']),
-                                new LinkTo('Progress bars', ['href'=>'#progress-bar'])
+                                new LinkTo('Progress bars', ['href'=>'#progress-bar']),
+                                new LinkTo('Thumbnails', ['href'=>'#thumbnails'])
                             ];
                         });
                         ?>