+++ src/Helpers/ContentTag.php
@@ -29,6 +29,7 @@
 namespace cobisja\BootHelp\Helpers;
 
 use cobisja\BootHelp\Base;
+use cobisja\BootHelp\Helpers\Badge;
 
 /**
  * ContentTag class: generates an specific HTML block tag surrounding an specific content.
@@ -67,8 +68,19 @@ class ContentTag extends Base {
      * @param mixed $options options to build the html object.
      */
     private function build_content_tag($name, $content, $options) {
-        ('a' === $name) && $this->get_alert_link() && $this->append_class($options, 'alert-link');
-        $content = is_object($content) ? $content->get_html_object() : $content;
-        $this->set_html_object($name, $options, $content);
+        $content = is_object($content) ? $content->get_html_object() : $content;        
+        $badge = $this->check_for_badge($options);        
+        $this->set_html_object($name, $options, is_null($badge) ? $content : [$content, Base::SPACE, $badge]);
     }
+    
+    /**
+     * Tells if badge options has been detected to generate a Badge component.
+     * 
+     * @param mixed $options options.
+     * @return Badge a instance of Badge component.
+     */
+    private function check_for_badge(&$options) {
+        $badge = Base::get_and_unset('badge', $options);        
+        return !is_null($badge) ? new Badge($badge) : null;
+    }    
 }