+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+**0.4.0** (Tuesday, 12th May 2015)
+
+* Added support for labels, images and thumbnails components.
+* Refactoring of few Base methods.
+* Guide updating.
+* Fix minor bugs.
+
 **0.3.2** (Friday, 1st May 2015)
 
 * Fix minor bug in progress bar label.